requirements.txt
Django>=2.1,<2.2
Pillow>=5.2,<5.3
requests>=2.19.0,<2.20.0
uWSGI==2.0.17.1
psycopg2-binary==2.7.5
django-debug-toolbar==1.9.1
# Version 0.6.21 of django-admin-sortable2 doesn't show add new row button in admin. New version is not released
git+git://github.com/jrief/django-admin-sortable2.git
django-allauth==0.38.0
django-reversion==3.0.0
gunicorn>=19.9,<19.10
ipdb==0.11

djangorestframework==3.8.2
django-rest-auth==0.9.3
django-filter==2.0.0
flex==6.13.2
drf-yasg==1.10.0
# This has a fix for migrations error
git+git://github.com/aschn/drf-tracking.git
firebase-admin==2.13.0

django-enumfields==0.10.1
django-redis==4.9.0
# Workaround for https://github.com/niwinz/django-redis/issues/342
redis==2.10.6
django-ipware==2.1.0
raven==6.9.0
# Version in pip doesn't have EXIF bugfix
git+git://github.com/un1t/django-resized.git
django-crispy-forms==1.7.2
django-recaptcha==1.4.0
python-dateutil==2.7.3
